[MyContacts Front-end] [Melhorando a Performance da Aplicação] - Melhorando as validações do Back-end

+++ api/src/app/controllers/CategoryController.js
@@ -1,4 +1,5 @@
 const CategoriesRepository = require("../repositories/CategoriesRepository");
+const isValidUUID = require("../utils/isValidUUID");
 
 class CategoryController {
   async index(request, response) {
@@ -10,6 +11,11 @@ class CategoryController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid category id" });
+    }
+
     const category = await CategoriesRepository.findById(id);
 
     if (!category) {
@@ -47,6 +53,10 @@ class CategoryController {
       return response.status(400).json({ error: "Name is required" });
     }
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid category id" });
+    }
+
     const categoryExists = await CategoriesRepository.findById(id);
 
     if (!categoryExists) {
@@ -68,6 +78,10 @@ class CategoryController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid category id" });
+    }
 
     await CategoriesRepository.delete(id);
